+++ lib/widgets/vs_node.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:vs_node_view/data/vs_node_data.dart';
-import 'package:vs_node_view/widgets/inherited_node_data_provider.dart';
+import 'package:vs_node_view/data/vs_node_data_provider.dart';
 import 'package:vs_node_view/widgets/vs_node_input.dart';
 import 'package:vs_node_view/widgets/vs_node_output.dart';
 import 'package:vs_node_view/widgets/vs_node_title.dart';
@@ -57,7 +57,7 @@ class _VSNodeState extends State<VSNode> {
       );
     }
 
-    final nodeProvider = InheritedNodeDataProvider.of(context).provider;
+    final nodeProvider = VSNodeDataProvider.of(context);
 
     return Draggable(
       onDragEnd: (_) {